dev to master

+++ src/main/java/com/rezztoran/rezztoranbe/service/impl/ReviewServiceImpl.java
@@ -2,8 +2,11 @@ package com.rezztoran.rezztoranbe.service.impl;
 
 import static java.util.stream.Collectors.groupingBy;
 
+import com.rezztoran.rezztoranbe.aop.AuthorizeCheck;
 import com.rezztoran.rezztoranbe.dto.ReviewDTO;
 import com.rezztoran.rezztoranbe.dto.request.ReviewRequestModel;
+import com.rezztoran.rezztoranbe.exception.BusinessException.Ex;
+import com.rezztoran.rezztoranbe.exception.ExceptionUtil;
 import com.rezztoran.rezztoranbe.model.Review;
 import com.rezztoran.rezztoranbe.repository.ReviewRepository;
 import com.rezztoran.rezztoranbe.service.RestaurantService;
@@ -26,6 +29,8 @@ public class ReviewServiceImpl implements ReviewService {
   private final UserService userService;
   private final RestaurantService restaurantService;
   private final ReviewRepository reviewRepository;
+  private final ExceptionUtil exceptionUtil;
+  private final AuthServiceImpl authService;
 
   private static ReviewDTO getReviewDTO(Review x) {
     return ReviewDTO.builder()
@@ -40,9 +45,17 @@ public class ReviewServiceImpl implements ReviewService {
   }
 
   @Override
+  @AuthorizeCheck(
+      field = "userId",
+      exceptRoles = {"ADMIN", "RESTAURANT_ADMIN"})
   public ReviewDTO createReview(ReviewRequestModel request) {
     var user = userService.findUserByID(request.getUserId());
     var restaurant = restaurantService.getById(request.getRestaurantId());
+
+    if (Boolean.TRUE.equals(
+        reviewRepository.existsByUser_IdAndRestaurant_Id(user.getId(), restaurant.getId()))) {
+      throw exceptionUtil.buildException(Ex.REVIEW_ALREADY_EXISTS_EXCEPTION);
+    }
 
     var savedRecord =
         reviewRepository.save(
@@ -78,12 +91,16 @@ public class ReviewServiceImpl implements ReviewService {
   public Review getReviewById(Long id) {
     return reviewRepository
         .findById(id)
-        .orElseThrow(() -> new RuntimeException("review not found!"));
+        .orElseThrow(() -> exceptionUtil.buildException(Ex.REVIEW_NOT_FOUND_EXCEPTION));
   }
 
   @Override
   public void deleteReviewById(Long id) {
+    var authUserId = authService.getAuthenticatedUser().get().getId();
     var existing = getReviewById(id);
+    if (!authUserId.equals(existing.getUser().getId())) {
+      throw exceptionUtil.buildException(Ex.FORBIDDEN_EXCEPTION);
+    }
     reviewRepository.delete(existing);
   }
